refactor, commented out, fixed
