+++ time_series/test_time_series/config/priors/default/species.ini
@@ -1,3 +1,4 @@
 [Species]
 growth_rate = u,0,1
 interactions = n
+observables = n
